+++ src/parser/mod.rs
@@ -1,2 +1,3 @@
+pub mod diff;
 pub mod json;
 pub mod zip;